55: extract `<OwnVocabTreasury>` into the newly-created `frontend/src/features/examples/` directory
