+++ frontend/src/App.test.tsx
@@ -20,21 +20,13 @@ import {
   Home,
   About,
   Account,
-  OwnVocabTreasury,
   RecordNewExample,
   SingleExample,
 } from "./App";
 import { Alerts } from "./features/alerts/Alerts";
 
 // 2
-import {
-  DefaultRequestBody,
-  RequestParams,
-  ResponseComposition,
-  rest,
-  RestContext,
-  RestRequest,
-} from "msw";
+import { rest } from "msw";
 import { setupServer } from "msw/node";
 
 import { applyMiddleware } from "redux";
@@ -57,7 +49,7 @@ import {
   IExampleFromBackend,
   IPaginationMeta,
   IExample,
-} from "./store";
+} from "./types";
 import { convertToPaginationInFrontend } from "./helperFunctionsForTesting";
 
 // 5
@@ -244,186 +236,6 @@ describe("<Account>", () => {
       expect(tableCellElement).toBeInTheDocument();
     }
   });
-});
-
-describe("<OwnVocabTreasury> + mocking of HTTP requests to the backend", () => {
-  beforeAll(() => {
-    /* Enable API mocking. */
-    quasiServer.listen();
-  });
-
-  afterEach(() => {
-    quasiServer.resetHandlers();
-  });
-
-  afterAll(() => {
-    /* Disable API mocking. */
-    quasiServer.close();
-  });
-
-  test(
-    "renders a heading, manipulation links," +
-      " and a page of the logged-in user's Example resources",
-    async () => {
-      /* Arrange. */
-      const initState: IState = {
-        ...INITIAL_STATE,
-        auth: {
-          ...INITIAL_STATE.auth,
-          loggedInUserProfile: {
-            id: 17,
-            username: "auth-jd",
-            email: "auth-john.doe@protonmail.com",
-          },
-        },
-      };
-      const enhancer = applyMiddleware(thunkMiddleware);
-      const realStore = createStore(rootReducer, initState, enhancer);
-
-      const history = createMemoryHistory();
-
-      const rhf: RequestHandlingFacilitator = new RequestHandlingFacilitator();
-      quasiServer.use(rest.get("/api/examples", rhf.createMockFetchExamples()));
-
-      /* Act. */
-      render(
-        <Provider store={realStore}>
-          <Router history={history}>
-            <OwnVocabTreasury />
-          </Router>
-        </Provider>
-      );
-
-      /* Assert. */
-      /* - after the component's rendering: */
-      /*    (a) items that appear above the table of Example resources */
-      let element: HTMLElement;
-      for (const expectedText of [
-        "auth-jd's Own VocabTreasury",
-        "Record new example",
-        "Search",
-      ]) {
-        element = screen.getByText(expectedText);
-        expect(element).toBeInTheDocument();
-      }
-
-      /*    (b) elements for controlling pagination of Example resources */
-      console.log(`${new Date().toISOString()} - 1st run of screen.debug()`);
-      screen.debug();
-
-      element = screen.getByText("Building pagination-controlling buttons...");
-      expect(element).toBeInTheDocument();
-
-      /*    (c) table headers */
-      for (const columnName of [
-        "ID",
-        "SOURCE LANGUAGE",
-        "NEW WORD",
-        "EXAMPLE",
-        "TRANSLATION",
-      ]) {
-        const tableCellElement = screen.getByText(columnName);
-        expect(tableCellElement).toBeInTheDocument();
-      }
-
-      /* - after the component's re-rendering: */
-      /*    (a) [representations of] Example resources */
-      element = await screen.findByText("1");
-
-      console.log(`${new Date().toISOString()} - 2nd run of screen.debug()`);
-      screen.debug();
-
-      expect(element).toBeInTheDocument();
-
-      for (const textFromExample1 of [
-        "sana numero-1",
-        "lause numero-1",
-        "käännös numero-1",
-      ]) {
-        element = screen.getByText(textFromExample1);
-        expect(element).toBeInTheDocument();
-      }
-
-      for (const textFromExample2 of [
-        "2",
-        "sana numero-2",
-        "lause numero-2",
-        "käännös numero-2",
-      ]) {
-        element = screen.getByText(textFromExample2);
-        expect(element).toBeInTheDocument();
-      }
-
-      /*    (b) elements for controlling pagination of Example resources */
-      for (const textOnPaginationControllingButton of [
-        "First page: 1",
-        "Previous page",
-        "Next page",
-        "Last page: 6",
-      ]) {
-        console.log("looking for ", textOnPaginationControllingButton);
-        element = screen.getByRole("button", {
-          name: textOnPaginationControllingButton,
-        });
-        expect(element).toBeInTheDocument();
-      }
-
-      element = screen.getByText("Current page: 1");
-      expect(element).toBeInTheDocument();
-    }
-  );
-
-  test(
-    " + <Alerts>" +
-      " a GET request to /api/examples is issued as part of the effect function," +
-      " but the backend is _mocked_" +
-      " to reject the client-provided authentication credential as invalid",
-    async () => {
-      /*
-      TODO: find out whether it would be better practice to convert this test case
-            into one that (fits under the
-            "multiple components + mocking of HTTP requests to the backend" `describe`
-            block and) renders the whole <App>, makes analogous assertions as this test,
-            and finally concludes by making the following extra assertion:
-              ```
-              expect(history.location.pathname).toEqual("/login")
-              ```
-      */
-
-      /* Arrange. */
-      const initState: IState = {
-        ...INITIAL_STATE,
-        auth: {
-          ...INITIAL_STATE.auth,
-          loggedInUserProfile: {
-            id: 17,
-            username: "auth-jd",
-            email: "auth-john.doe@protonmail.com",
-          },
-        },
-      };
-      const enhancer = applyMiddleware(thunkMiddleware);
-      const realStore = createStore(rootReducer, initState, enhancer);
-
-      const history = createMemoryHistory();
-
-      /* Act. */
-      render(
-        <Provider store={realStore}>
-          <Router history={history}>
-            <Alerts />
-            <OwnVocabTreasury />
-          </Router>
-        </Provider>
-      );
-
-      /* Assert. */
-      const element: HTMLElement = await screen.findByText(
-        "[FROM <OwnVocabTreasury>'S useEffect HOOK] PLEASE LOG BACK IN"
-      );
-      expect(element).toBeInTheDocument();
-    }
-  );
 });
 
 describe("<RecordNewExample>", () => {
