+++ frontend/src/constants.ts
@@ -14,3 +14,20 @@ export const INITIAL_STATE_AUTH: IStateAuth = {
   hasValidToken: null,
   loggedInUserProfile: null,
 };
+
+export const URL_FOR_FIRST_PAGE_OF_EXAMPLES: string = "/api/examples";
+
+/*
+TODO: (2023/10/29, 11:52)
+
+      before submitting a pull request for review,
+      rename the following variables so that they will "read" like constants
+
+      additional follow-up:
+      consider whether those variables will be rendered unused/unnecessary
+      when v-t-i-58 has been resolved
+*/
+export const styleForBorder = { border: "1px solid black" };
+
+export const styleForTable = { width: "100%" };
+Object.assign(styleForTable, styleForBorder);
